requirements.txt
annotated-types==0.7.0
anyio==3.7.1
Authlib==1.3.1
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cryptography==45.0.4
deprecation==2.1.0
exceptiongroup==1.3.0
fastapi==0.104.1
filelock==3.18.0
fsspec==2025.5.1
greenlet==3.2.3
grpcio==1.73.0
grpcio-health-checking==1.73.0
grpcio-tools==1.73.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.33.0
idna==3.10
Jinja2==3.1.6
joblib==1.5.1
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
nltk==3.9.1
numpy==1.24.3
packaging==25.0
pillow==11.2.1
protobuf==6.31.1
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.11.4
sentence-transformers==4.1.0
sentencepiece==0.2.0
sniffio==1.3.1
SQLAlchemy==2.0.23
starlette==0.27.0
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
transformers==4.52.4
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.5.0
uvicorn==0.24.0
validators==0.34.0
watchfiles==1.1.0
weaviate-client==4.15.2
websockets==15.0.1
